+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `v0.0.14` Form component
   - Fix bugs where query data was always merged with form data when updated by adding a new `shouldMergeQueryData` modifier, set to `true` by default.
   - Fix visual bug where readonly form table cells would sometimes have height too large
+  - Fixed an oversight where custom fields would not display their error tooltip in read-mode
 
 ## 0.33.12 - 2024-10-29
 ### Fixed